requirements.txt
groq
gradio
python-dotenv
chromadb
sentence-transformers
pypdf
requests  